services/data-generator/requirements.txt
kafka-python==2.0.2
faker==22.0.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
